backend/structura_backend/requirements.txt
Django>=5.2,<6
djangorestframework>=3.15
django-cors-headers>=4.6
psycopg2-binary>=2.9
Gunicorn>=21.2
whitenoise>=6.7
dj-database-url>=2.2
python-dotenv>=1.0
mediapipe
Pillow
numpy
